Member Highlight
This script was written to  displays the picture and bio of one member
of the organization on the organizations website and changes it once a
week.  The script determines the uses an algorithm to ensure that
everyone is highlighted at least once before repeating any members. It
increments the highlight count by one when a member is highlighted.
It only displays a small sample of the bio, for a full bio the user will
click on the link to view the members full page.